requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==33.1.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Django==4.2.12
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1